Split tests into separate modules
