+++ tests/test_radiation.py
@@ -1,0 +1,4 @@
+"""
+Tests for prepic.radiation module
+
+"""
